2021/18 Alternative Implementation
I tried to do the tree model last night
and gave up when I hit "nearest left"
and "nearest right". parent pointers
allow me to solve that, so I wanted to
flesh out the idea.

The stack implementation is _way_
faster howerver, so keep it and  add
an audit flag to allow testing to hold it
around, but not run it in benchmarks.