2021/19 Problem Assertion
This solution leverages the indication
from the problem that 12 beacons of
overlap is both necessary and
sufficient to align scanners.

Add CHECKs that when we do set up
an alignment all of the becaons that
don't line up are in fact not visible
from the scanners in which they are
not present.