2020/13: Use Chinese Remainder Theorem
I implemented the Chinese Remainder model for handling systems of
modular equations for project euler, and this is backporting it to an
existing AOC problem. I was hoping that this would be a performance
gain, but generally it's worse on the test input, with the real input
being 5us instead of 10us.