Add a (very simplisitic) OCR method
For the cases where we need to return a message encoded in a CharBoard,
we've leaned on a human being able to do the final translation to what
is input to the website. This new code can extract the requisite letters
from those boards into a single string that can be directly copy&pasted.