2016/11 Performance
Down to ~16s from 1m. Primarily by reducing the size of Floor and
inlining it. So we now use char for the bit vector, and we don't drag
along the elements vector (but rather a pointer to it).

This also wraps Floor up as a class with an interface, so the details
aren't exposed out and the process of creating new states in the
BFSInterface is a little more readable.