Update 2016/11 Parser
I tested out get_input with 2016/11 and apparently the contents of my
input has changed since I originally solved the problem. There is a
slight different in formatting a pair of elements on a floor and I've
updated the parser to handle both.

I've updated the example tests to include both forms.