TODONE: 2017/20
Rather than run a simulation with a hard-coded stop value, we do a
pairwise, and comprehensive, intersection of particles. This is slightly
slower (since we're doing a more complicated N^2 with N ~ the hard
coded step limit), but it's still <1s and I like not having to have the
magic step limit.