New option: [flags=ocr_mode={input,output,both}]
This creates a flag= option that sets an absl command line flag within a
test. We then use this option to set a new ocr_mode flag which controls
whether the test output should be the raw CharBoard, the result of OCR
or (the default, which is what --config=compete should do) both.